requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
pathlib>=1.0.1
typing-extensions>=4.9.0 